feat: Assign book to new shelf, filter view
Changed MoveToGrid to FilterBooks; filtered view in this component
and passed moveToShelf function from this component down the tree.
Had to initiate "view filters" in DisplayShelf and SearchAllBooks
which feed into FilterBooks -- this may not be best practice for
pure functions? Finally changed my draft App to the main App, and
renamed the starter code "App" to App-old. Nothing is tested yet.
Implemented EscapeRegExp in SearchAllBooks, echoing the React
"Contacts" project from before, but I may need to install it.