refactor: Separate filter functions from grid UI
Revise FilterBooks to reflect its name as a filter function for both
shelves and search -- no longer a component in itself, rather imported
into RenderGrid (separating the interface from the logic). My intent is
to clarify code and intent, and to allow better isolation as I test for
bugs that continue to arise in my search implementation.

Also fixed hoisting of the "value" variable for a book's shelf state.
As previously implemented, it was not passed into the search API (this
reflects my previous implementation of "search" only within the user
library, not all books available within the API).

And removed some console.log statements previously used for testing.